[CHORE] : organized code a bit
