+++ frontend/components/main/member_list/MemberList.tsx
@@ -1,20 +1,14 @@
 "use client";
 
 import "@/components/main/room_list/RoomList.css";
+import "@/components/main/member_list/MemberList.css";
 import Title from "../room_list/Title";
 import Members from "./Members";
-import { useEffect } from "react";
-import "@/components/main/member_list/MemberList.css";
 import { useRoom } from "@/context/RoomContext";
 
-export default function MemberList({
-}: {
-}) {
+export default function MemberList({}: {}) {
   const { roomState } = useRoom();
 
-  // useEffect(() => {
-  //   isRight ? setIsRight(false) : null;
-  // }, [isRight]);
   return (
     <>
       {roomState.isOpen ? (
