+++ frontend/components/main/room_list/Room.tsx
@@ -1,17 +1,42 @@
 "use client";
 
-import LockRoundedIcon from "@mui/icons-material/LockRounded";
 import { useEffect, useState, useRef } from "react";
+import LockRoundedIcon from "@mui/icons-material/LockRounded";
 import ProtectedModal from "./ProtectedModal";
-import { IChatEnter, useRoom } from "@/context/RoomContext";
+import { IChatEnter, IChatEnterNoti, alert, useRoom } from "@/context/RoomContext";
 import { IChatRoom0, Mode } from "@/context/RoomContext";
 import { socket } from "@/app/layout";
+import Alert from "@mui/material/Alert";
 
 export default function Room({ room, idx }: { room: IChatRoom0; idx: number }) {
   const [open, setOpen] = useState(false);
   const [fail, setFail] = useState<boolean>(false);
-  const statusCode = useRef(0);
+  const [showAlert, setShowAlert] = useState<boolean>(false);
+  const [newMem, setNewMem] = useState("");
   const { roomState, roomDispatch } = useRoom();
+
+  useEffect(() => {
+    const ChatEnterNoti = (data: IChatEnterNoti) => {
+      //alert
+      setShowAlert(true);
+      setNewMem(data.newMember);
+    };
+    socket.on("chat_enter_noti", ChatEnterNoti);
+
+    return () => {
+      socket.off("chat_enter_noti", ChatEnterNoti);
+    };
+  }); // TODO : 위치?
+
+  useEffect(() => {
+    if (showAlert) {
+      const time = setTimeout(() => {
+        setShowAlert(false);
+      }, 3000);
+
+      return () => clearTimeout(time);
+    }
+  }, [showAlert]);
 
   const leftPadding = (idx: number) => {
     if (idx < 10) return "00" + idx.toString();
@@ -33,7 +58,6 @@ export default function Room({ room, idx }: { room: IChatRoom0; idx: number }) {
 
   useEffect(() => {
     const ChatEnter = (json: IChatEnter) => {
-      console.log("ChatEnter : ", json);
       roomDispatch({ type: "SET_CUR_MEM", value: json.member });
       //channelIdx 안보내줘도 될듯?
     };
@@ -46,7 +70,6 @@ export default function Room({ room, idx }: { room: IChatRoom0; idx: number }) {
 
   useEffect(() => {
     const ChatDmEnter = (json: any) => {
-      console.log("ChatDmEnter : ", json);
       roomDispatch({
         type: "SET_CUR_DM_MEM",
         value: {
@@ -74,9 +97,7 @@ export default function Room({ room, idx }: { room: IChatRoom0; idx: number }) {
             channelIdx: room.channelIdx,
           }),
           (json: any) => {
-            console.log("chat_get_DM : ", json);
             if (roomState.currentRoom !== room) {
-              console.log("thissss");
               roomDispatch({ type: "SET_CURRENTROOM", value: room });
             }
             roomDispatch({ type: "SET_ISOPEN", value: true });
@@ -107,21 +128,19 @@ export default function Room({ room, idx }: { room: IChatRoom0; idx: number }) {
 
   return (
     <>
-      <>
-        <button key={idx} className="item" onClick={() => RoomClick(room)}>
-          <div className="roomidx">{leftPadding(room.channelIdx)}</div>
-          <div className="owner">
-            {room.owner ? room.owner : room.targetNickname}'s
-          </div>
-          <div className="lock">
-            {room.mode === Mode.PROTECTED ? (
-              <LockRoundedIcon sx={{ height: "13px", color: "#afb2b3" }} />
-            ) : (
-              ""
-            )}
-          </div>
-        </button>
-      </>
+      <button key={idx} className="item" onClick={() => RoomClick(room)}>
+        <div className="roomidx">{leftPadding(room.channelIdx)}</div>
+        <div className="owner">
+          {room.owner ? room.owner : room.targetNickname}'s
+        </div>
+        <div className="lock">
+          {room.mode === Mode.PROTECTED ? (
+            <LockRoundedIcon sx={{ height: "13px", color: "#afb2b3" }} />
+          ) : (
+            ""
+          )}
+        </div>
+      </button>
       <ProtectedModal
         open={open}
         handleClose={handleClose}
@@ -129,6 +148,11 @@ export default function Room({ room, idx }: { room: IChatRoom0; idx: number }) {
         fail={fail}
         setFail={setFail}
       />
+      {showAlert ? (
+          <Alert sx={alert} severity="info" style={{ width: "333px" }}>
+            {newMem} has joined
+          </Alert>
+        ) : null}
     </>
   );
 }
