+++ frontend/components/main/room_list/Rooms.tsx
@@ -7,24 +7,18 @@ import MemberList from "../member_list/MemberList";
 import CreateRoomButton from "./CreateRoomButton";
 import Room from "./Room";
 import { IChatRoom0, useRoom } from "@/context/RoomContext";
-// import { IChatRoom0, IDmRoom, useRoom } from "@/context/RoomContext";
 
 export default function Rooms({
   currentRoomList,
   channelType,
 }: {
   currentRoomList: IChatRoom0[];
-  // currentRoomList: IChatRoom0[] | IDmRoom[];
   channelType: boolean;
 }) {
   const [portalContainer, setPortalContainer] = useState<HTMLElement | null>(
     null
   );
   const { roomState } = useRoom();
-
-  // useEffect(() => {
-  //   isRight ? setIsOpen(true) : null;
-  // }, [isRight]);
 
   useEffect(() => {
     const container = document.getElementById("portal");
