+++ frontend/components/public/Layout.tsx
@@ -12,18 +12,10 @@ import InviteGame from "../main/InviteGame/InviteGame";
 import WaitAccept from "../main/InviteGame/WaitAccept";
 import { useEffect, useState } from "react";
 import { useAuth } from "@/context/AuthContext";
-import { IChatEnterNoti, useRoom } from "@/context/RoomContext";
+import { useRoom } from "@/context/RoomContext";
 import { UserProvider, useUser } from "@/context/UserContext";
 import { socket } from "@/app/layout";
 import { useFriend } from "@/context/FriendContext";
-import Alert from "@mui/material/Alert";
-
-const alert = {
-  position: "absolute" as "absolute",
-  top: "100%",
-  left: "50%",
-  transform: "translate(-50%, -100%)",
-};
 
 export const main = {
   main0: "#67DBFB",
@@ -82,8 +74,6 @@ const Layout = () => {
   const { roomState, roomDispatch } = useRoom();
   const { friendState, friendDispatch } = useFriend();
   const { userState, userDispatch } = useUser();
-  const [showAlert, setShowAlert] = useState<boolean>(false);
-  const [newMem, setNewMem] = useState("");
 
   useRequireAuth();
 
@@ -108,35 +98,11 @@ const Layout = () => {
     };
   }, []);
 
-  useEffect(() => {
-    const ChatEnterNoti = (data: IChatEnterNoti) => {
-      console.log("data : ", data);
-      //alert
-      setShowAlert(true);
-      setNewMem(data.newMember);
-    };
-    socket.on("chat_enter_noti", ChatEnterNoti);
-
-    return () => {
-      socket.off("chat_enter_noti", ChatEnterNoti);
-    };
-  }); // TODO : 위치?
-
-  useEffect(() => {
-    if (showAlert) {
-      const time = setTimeout(() => {
-        setShowAlert(false);
-      }, 3000);
-
-      return () => clearTimeout(time);
-    }
-  }, [showAlert]);
-
   useRequireAuth();
 
   useEffect(() => {
     if (state.isLoggedIn) {
-      socket.emit("main_enter", JSON.stringify({ intra: "jeekim" }), () => {});
+      socket.emit("main_enter", JSON.stringify({ intra: "hoslim" }), () => {});
     }
   }, [state.isLoggedIn]);
 
@@ -191,11 +157,7 @@ const Layout = () => {
         }}
       >
         <RoomList />
-        {showAlert ? (
-          <Alert sx={alert} severity="info" style={{ width: "333px" }}>
-            {newMem} has joined
-          </Alert>
-        ) : null}
+        
       </Stack>
     </Box>
   );
